+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix name and namespace regex used in create project from example command
+
 ## [1.2.1] 2022-12-12
 
 ### Fixed